+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 0.0.3
 
+* `Common`:
+    * Fix in initialization of `InFileStartedTimersRepo`
+
 ## 0.0.2
 
 * Dependencies: